Split configuration into separate files
